+++ Tests/SwiftStdlibTests/SequenceExtensionsTests.swift
@@ -26,4 +26,64 @@ final class SequenceExtensionsTests: XCTestCase {
         XCTAssert(collection.none { $0 % 2 == 0 })
     }
 
+    func testLastWhere() {
+        let array = [1, 1, 2, 1, 1, 1, 2, 1, 4, 1]
+        let element = array.last { $0 % 2 == 0 }
+        XCTAssertEqual(element, 4)
+        XCTAssertNil([Int]().last { $0 % 2 == 0 })
+    }
+
+    func testRejectWhere() {
+        let input = [1, 2, 3, 4, 5]
+        let output = input.reject { $0 % 2 == 0 }
+        XCTAssertEqual(output, [1, 3, 5])
+    }
+
+    func testCountWhere() {
+        let array = [1, 1, 1, 1, 4, 4, 1, 1, 1]
+        let count = array.count { $0 % 2 == 0 }
+        XCTAssertEqual(count, 2)
+    }
+
+    func testForEachReversed() {
+        let input = [1, 2, 3, 4, 5]
+        var output: [Int] = []
+        input.forEachReversed { output.append($0) }
+        XCTAssertEqual(output.first, 5)
+    }
+
+    func testForEachWhere() {
+        let input = [1, 2, 2, 2, 1, 4, 1]
+        var output: [Int] = []
+        input.forEach(where: {$0 % 2 == 0}, body: { output.append($0 * 2) })
+        XCTAssertEqual(output, [4, 4, 4, 8])
+    }
+
+    func testAccumulate() {
+        let input = [1, 2, 3]
+        let result = input.accumulate(initial: 0, next: +)
+        XCTAssertEqual([1, 3, 6], result)
+    }
+
+    func testFilteredMap() {
+        let input = [1, 2, 3, 4, 5]
+        let result = input.filtered({ $0 % 2 == 0 }, map: { $0.string })
+        XCTAssertEqual(result.count, 2)
+        XCTAssertEqual(["2", "4"], result)
+    }
+
+    func testContains() {
+        XCTAssert([Int]().contains([]))
+        XCTAssertFalse([Int]().contains([1, 2]))
+        XCTAssert([1, 2, 3].contains([1, 2]))
+        XCTAssert([1, 2, 3].contains([2, 3]))
+        XCTAssert([1, 2, 3].contains([1, 3]))
+        XCTAssertFalse([1, 2, 3].contains([4, 5]))
+    }
+
+    func testSum() {
+        XCTAssertEqual([1, 2, 3, 4, 5].sum(), 15)
+        XCTAssertEqual([1.2, 2.3, 3.4, 4.5, 5.6].sum(), 17)
+    }
+
 }